feat(modules): migrate from namespace to ES modules
Convert all TypeScript files from namespace pattern to ES modules:
- Add export statements to all classes, interfaces, and functions
- Add import statements where dependencies are needed
- Remove namespace declarations from project.ts and project-state.ts
- Update all component files with proper ES module syntax

Files converted:
- decorators/autobind.ts
- util/validation.ts
- models/project.ts (namespace removed)
- components/base-component.ts
- state/project-state.ts (namespace removed)
- components/project-item.ts
- components/project-input.ts
- components/project-list.ts

All TypeScript compilation errors resolved. The project now uses modern ES2015 modules as configured in tsconfig.json.
