+++ scripts/ensure-rx-entrypoint.js
