+++ src/app.ts
@@ -1,389 +1,8 @@
-//D&D
-interface Draggable {
-  dragStartHandler(evt: DragEvent): void;
-  dragEndHandler(event: DragEvent): void;
+import { ProjectInput } from './components/project-input.js';
+import { ProjectList } from './components/project-list.js';
+
+namespace App {
+  new ProjectInput();
+  new ProjectList('active');
+  new ProjectList('finished');
 }
-
-interface DragTarget {
-  dragOverHandler(event: DragEvent): void;
-  dropHandler(evt: DragEvent): void;
-  dragLeaveHandler(evt: DragEvent): void;
-}
-
-enum ProjectStatus {
-  active,
-  finished,
-}
-
-//型をクラスで定義？？何でここはインスタンスを作りたい？
-class Project {
-  constructor(
-    public id: string,
-    public title: string,
-    public description: string,
-    public manday: number,
-    public status: ProjectStatus
-  ) {}
-}
-
-//Project State Manegement
-type Listener<T> = (item: T[]) => void;
-
-class State<T> {
-  protected listeners: Listener<T>[] = [];
-
-  addListener(addListener: Listener<T>) {
-    this.listeners.push(addListener);
-  }
-}
-
-//シングルトンのクラス　全体でインスタンスが一つ？
-class ProjectState extends State<Project> {
-  private projects: Project[] = [];
-  private static instance: ProjectState;
-
-  private constructor() {
-    super();
-  }
-
-  static getInstance() {
-    if (this.instance) {
-      return this.instance;
-    }
-    this.instance = new ProjectState();
-    return this.instance;
-  }
-
-  addProject(title: string, description: string, manday: number) {
-    const newProject = new Project(
-      Math.random().toString(),
-      title,
-      description,
-      manday,
-      ProjectStatus.active
-    );
-    this.projects.push(newProject);
-    for (const listenersFn of this.listeners) {
-      listenersFn(this.projects.slice());
-    } //ここではオリジナルでなくコピーの配列を渡している？
-  }
-
-  moveProject(prjId: string, newStatus: ProjectStatus) {
-    const project = this.projects.find((prj) => prj.id === prjId);
-    if (project && project.status !== newStatus) {
-      project.status = newStatus;
-      this.updateListeners();
-    }
-  }
-
-  private updateListeners() {
-    for (const listenersFn of this.listeners) {
-      listenersFn(this.projects.slice());
-    }
-  }
-}
-
-const projectState = ProjectState.getInstance();
-
-//Validation
-//１．オブジェクトの型定義
-interface Validatable {
-  value: string | number;
-  required: boolean | undefined; //? と | undefined はおなじ
-  minLength?: number;
-  maxLength?: number;
-  min?: number;
-  max?: number;
-}
-
-//validateの関数定義
-function validate(validatableInput: Validatable) {
-  let isValid = true; //これの必要性は？
-  if (validatableInput.required) {
-    isValid = isValid && validatableInput.value.toString().trim().length !== 0;
-  }
-  if (
-    validatableInput.minLength != null &&
-    typeof validatableInput.value === 'string'
-  ) {
-    isValid =
-      isValid && validatableInput.value.length >= validatableInput.minLength;
-  }
-  if (
-    validatableInput.maxLength != null &&
-    typeof validatableInput.value === 'string'
-  ) {
-    isValid =
-      isValid && validatableInput.value.length <= validatableInput.maxLength;
-  }
-  if (
-    validatableInput.min != null &&
-    typeof validatableInput.value === 'number'
-  ) {
-    isValid = isValid && validatableInput.value >= validatableInput.min;
-  }
-  if (
-    validatableInput.max != null &&
-    typeof validatableInput.value === 'number'
-  ) {
-    isValid = isValid && validatableInput.value <= validatableInput.max;
-  }
-
-  return isValid;
-}
-
-//decorators
-function Autobind(
-  target: any,
-  methodName: string,
-  descriptor: PropertyDescriptor
-) {
-  const originalMethod = descriptor.value;
-  const boundDescriptor: PropertyDescriptor = {
-    configurable: true,
-    enumerable: false,
-    get() {
-      return originalMethod.bind(this);
-    },
-  };
-  return boundDescriptor;
-}
-
-//component class これは常に継承されるべき
-abstract class Component<T extends HTMLElement, U extends HTMLElement> {
-  templateElement: HTMLTemplateElement;
-  hostElement: T;
-  element: U;
-
-  constructor(
-    templateId: string,
-    hostElementId: string,
-    insertAtStart: boolean,
-    newElementId?: string
-  ) {
-    this.templateElement = document.getElementById(
-      templateId
-    )! as HTMLTemplateElement;
-    this.hostElement = document.getElementById(hostElementId)! as T;
-
-    const importNode = document.importNode(this.templateElement.content, true);
-    this.element = importNode.firstElementChild as U;
-    if (newElementId) {
-      this.element.id = newElementId;
-    }
-    this.attach(insertAtStart);
-  }
-
-  abstract configure(): void;
-
-  abstract renderContent(): void;
-
-  private attach(insertAtBeginning: boolean) {
-    this.hostElement.insertAdjacentElement(
-      insertAtBeginning ? 'afterbegin' : 'beforeend',
-      this.element
-    );
-  }
-}
-
-class ProjectItem
-  extends Component<HTMLUListElement, HTMLLIElement>
-  implements Draggable
-{
-  private project: Project;
-
-  get manday() {
-    if (this.project.manday < 20) {
-      return this.project.manday.toString() + ' (light)';
-    } else {
-      return (this.project.manday / 20).toString() + ' (heavy)';
-    }
-  }
-
-  constructor(hostId: string, project: Project) {
-    super('single-project', hostId, false, project.id);
-    this.project = project;
-
-    this.configure();
-    this.renderContent();
-  }
-
-  @Autobind
-  dragStartHandler(evt: DragEvent): void {
-    evt.dataTransfer!.setData('text/plain', this.project.id);
-    evt.dataTransfer!.effectAllowed = 'move';
-  }
-
-  dragEndHandler(__: DragEvent): void {
-    console.log('DragEnd');
-  }
-
-  configure() {
-    this.element.addEventListener('dragstart', this.dragStartHandler);
-    this.element.addEventListener('dragend', this.dragEndHandler);
-  }
-
-  renderContent() {
-    this.element.querySelector('h2')!.textContent = this.project.title;
-    this.element.querySelector('h3')!.textContent = this.manday;
-    this.element.querySelector('p')!.textContent = this.project.description;
-  }
-}
-
-//ProjectList Class
-class ProjectList
-  extends Component<HTMLDivElement, HTMLElement>
-  implements DragTarget
-{
-  assignedProjects: Project[];
-
-  constructor(private type: 'active' | 'finished') {
-    super('project-list', 'app', false, `${type}-projects`); //superの呼び出しが終わるまでthisが使えない？
-
-    this.assignedProjects = [];
-
-    this.configure();
-    this.renderContent();
-  }
-
-  @Autobind
-  dragOverHandler(evt: DragEvent): void {
-    if (evt.dataTransfer && evt.dataTransfer.types[0] === 'text/plain') {
-      evt.preventDefault();
-      const listEl = this.element.querySelector('ul')!;
-      listEl.classList.add('droppable');
-    }
-  }
-
-  dropHandler(evt: DragEvent): void {
-    const prjId = evt.dataTransfer!.getData('text/palin');
-    projectState.moveProject(
-      prjId,
-      this.type === 'active' ? ProjectStatus.active : ProjectStatus.finished
-    );
-  }
-
-  dragLeaveHandler(_: DragEvent): void {
-    const listEl = this.element.querySelector('ul')!;
-    listEl.classList.remove('droppable');
-  }
-
-  configure() {
-    this.element.addEventListener('dragover', this.dragOverHandler);
-    this.element.addEventListener('drop', this.dropHandler);
-    this.element.addEventListener('dragleave', this.dragLeaveHandler);
-
-    projectState.addListener((projects: Project[]) => {
-      const relevantProjects = projects.filter((prj) => {
-        if (this.type === 'active') {
-          return prj.status === ProjectStatus.active;
-        }
-        return prj.status === ProjectStatus.finished;
-      });
-      this.assignedProjects = relevantProjects;
-      this.renderProjects();
-    });
-  }
-
-  renderContent() {
-    const listId = `${this.type}-project-list`;
-    this.element.querySelector('ul')!.id = listId;
-    this.element.querySelector('h2')!.textContent =
-      this.type === 'active' ? 'active projects' : 'finished projects';
-  }
-
-  private renderProjects() {
-    const listEl = document.getElementById(
-      `${this.type}-project-list`
-    )! as HTMLUListElement;
-    listEl.innerHTML = '';
-    for (const prjItem of this.assignedProjects) {
-      new ProjectItem(listEl.id, prjItem);
-    }
-  }
-}
-
-//Project Input class
-class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
-  titleInputElement: HTMLInputElement;
-  descriptionInputElement: HTMLInputElement;
-  mandayInputElement: HTMLInputElement;
-
-  //変数に格納する vs this.property これの判断基準は？
-  constructor() {
-    super('project-input', 'app', true, 'user-input');
-
-    this.titleInputElement = this.element.querySelector(
-      '#title'
-    ) as HTMLInputElement;
-    this.descriptionInputElement = this.element.querySelector(
-      '#description'
-    ) as HTMLInputElement;
-    this.mandayInputElement = this.element.querySelector(
-      '#manday'
-    ) as HTMLInputElement;
-
-    this.configure();
-  }
-
-  configure() {
-    this.element.addEventListener('submit', this.submitHandler); //何でここはthisが必要？
-  }
-
-  renderContent() {}
-
-  private gatherUserInput(): [string, string, number] | void {
-    const enteredTitle = this.titleInputElement.value;
-    const enteredDescription = this.descriptionInputElement.value;
-    const enteredManday = this.mandayInputElement.value;
-    const titleValidatable: Validatable = {
-      value: enteredTitle,
-      required: true,
-    };
-    const descriptionValidatable: Validatable = {
-      value: enteredDescription,
-      required: true,
-      minLength: 5,
-    };
-    const mandayValidatable: Validatable = {
-      value: enteredManday,
-      required: true,
-      min: 1,
-      max: 1000,
-    };
-
-    if (
-      !validate(titleValidatable) ||
-      !validate(descriptionValidatable) ||
-      !validate(mandayValidatable)
-    ) {
-      alert('Invalid input value');
-      return;
-    } else {
-      return [enteredTitle, enteredDescription, +enteredManday];
-    }
-  }
-
-  private clearUserInput() {
-    this.titleInputElement.value = '';
-    this.descriptionInputElement.value = '';
-    this.mandayInputElement.value = '';
-  }
-
-  @Autobind
-  private submitHandler(evt: Event) {
-    evt.preventDefault();
-    console.log(this.titleInputElement.value);
-    const userInput = this.gatherUserInput();
-    if (Array.isArray(userInput)) {
-      const [title, desc, manday] = userInput;
-      projectState.addProject(title, desc, manday);
-      console.log(title, desc, manday);
-      this.clearUserInput();
-    }
-  }
-}
-
-const projectInput = new ProjectInput();
-const activePrjList = new ProjectList('active');
-const finishedPrjList = new ProjectList('finished');
