+++ src/components/project-input.ts
@@ -1,0 +1,85 @@
+import { Component } from './base-component.js';
+import { Validatable, validate } from '../util/validation.js';
+import { Autobind } from '../decorators/autobind.js';
+import { projectState } from '../state/project-state.js';
+
+//Project Input class
+export class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
+  titleInputElement: HTMLInputElement;
+  descriptionInputElement: HTMLInputElement;
+  mandayInputElement: HTMLInputElement;
+
+  //変数に格納する vs this.property これの判断基準は？
+  constructor() {
+    super('project-input', 'app', true, 'user-input');
+
+    this.titleInputElement = this.element.querySelector(
+      '#title'
+    ) as HTMLInputElement;
+    this.descriptionInputElement = this.element.querySelector(
+      '#description'
+    ) as HTMLInputElement;
+    this.mandayInputElement = this.element.querySelector(
+      '#manday'
+    ) as HTMLInputElement;
+
+    this.configure();
+  }
+
+  configure() {
+    this.element.addEventListener('submit', this.submitHandler); //何でここはthisが必要？
+  }
+
+  renderContent() {}
+
+  private gatherUserInput(): [string, string, number] | void {
+    const enteredTitle = this.titleInputElement.value;
+    const enteredDescription = this.descriptionInputElement.value;
+    const enteredManday = this.mandayInputElement.value;
+    const titleValidatable: Validatable = {
+      value: enteredTitle,
+      required: true,
+    };
+    const descriptionValidatable: Validatable = {
+      value: enteredDescription,
+      required: true,
+      minLength: 5,
+    };
+    const mandayValidatable: Validatable = {
+      value: enteredManday,
+      required: true,
+      min: 1,
+      max: 1000,
+    };
+
+    if (
+      !validate(titleValidatable) ||
+      !validate(descriptionValidatable) ||
+      !validate(mandayValidatable)
+    ) {
+      alert('Invalid input value');
+      return;
+    } else {
+      return [enteredTitle, enteredDescription, +enteredManday];
+    }
+  }
+
+  private clearUserInput() {
+    this.titleInputElement.value = '';
+    this.descriptionInputElement.value = '';
+    this.mandayInputElement.value = '';
+  }
+
+  @Autobind
+  private submitHandler(evt: Event) {
+    evt.preventDefault();
+    console.log(this.titleInputElement.value);
+    const userInput = this.gatherUserInput();
+    if (Array.isArray(userInput)) {
+      const [title, desc, manday] = userInput;
+      projectState.addProject(title, desc, manday);
+      console.log(title, desc, manday);
+      this.clearUserInput();
+    }
+  }
+}
