+++ src/components/project-list.ts
@@ -1,0 +1,79 @@
+import { Component } from './base-component.js';
+import { DragTarget } from '../models/dd-interfaces.js';
+import { Project, ProjectStatus } from '../models/project.js';
+import { Autobind } from '../decorators/autobind.js';
+import { projectState } from '../state/project-state.js';
+import { ProjectItem } from './project-item.js';
+
+//ProjectList Class
+export class ProjectList
+  extends Component<HTMLDivElement, HTMLElement>
+  implements DragTarget
+{
+  assignedProjects: Project[];
+
+  constructor(private type: 'active' | 'finished') {
+    super('project-list', 'app', false, `${type}-projects`); //superの呼び出しが終わるまでthisが使えない？
+
+    this.assignedProjects = [];
+
+    this.configure();
+    this.renderContent();
+  }
+
+  @Autobind
+  dragOverHandler(evt: DragEvent): void {
+    if (evt.dataTransfer && evt.dataTransfer.types[0] === 'text/plain') {
+      evt.preventDefault();
+      const listEl = this.element.querySelector('ul')!;
+      listEl.classList.add('droppable');
+    }
+  }
+
+  dropHandler(evt: DragEvent): void {
+    const prjId = evt.dataTransfer!.getData('text/palin');
+    projectState.moveProject(
+      prjId,
+      this.type === 'active' ? ProjectStatus.active : ProjectStatus.finished
+    );
+  }
+
+  dragLeaveHandler(_: DragEvent): void {
+    const listEl = this.element.querySelector('ul')!;
+    listEl.classList.remove('droppable');
+  }
+
+  configure() {
+    this.element.addEventListener('dragover', this.dragOverHandler);
+    this.element.addEventListener('drop', this.dropHandler);
+    this.element.addEventListener('dragleave', this.dragLeaveHandler);
+
+    projectState.addListener((projects: Project[]) => {
+      const relevantProjects = projects.filter((prj) => {
+        if (this.type === 'active') {
+          return prj.status === ProjectStatus.active;
+        }
+        return prj.status === ProjectStatus.finished;
+      });
+      this.assignedProjects = relevantProjects;
+      this.renderProjects();
+    });
+  }
+
+  renderContent() {
+    const listId = `${this.type}-project-list`;
+    this.element.querySelector('ul')!.id = listId;
+    this.element.querySelector('h2')!.textContent =
+      this.type === 'active' ? 'active projects' : 'finished projects';
+  }
+
+  private renderProjects() {
+    const listEl = document.getElementById(
+      `${this.type}-project-list`
+    )! as HTMLUListElement;
+    listEl.innerHTML = '';
+    for (const prjItem of this.assignedProjects) {
+      new ProjectItem(listEl.id, prjItem);
+    }
+  }
+}
