+++ src/decorators/autobind.ts
@@ -1,0 +1,16 @@
+//decorators
+export function Autobind(
+  target: any,
+  methodName: string,
+  descriptor: PropertyDescriptor
+) {
+  const originalMethod = descriptor.value;
+  const boundDescriptor: PropertyDescriptor = {
+    configurable: true,
+    enumerable: false,
+    get() {
+      return originalMethod.bind(this);
+    },
+  };
+  return boundDescriptor;
+}
