+++ src/util/validation.ts
@@ -1,0 +1,46 @@
+//Validation
+//１．オブジェクトの型定義
+export interface Validatable {
+  value: string | number;
+  required: boolean | undefined; //? と | undefined はおなじ
+  minLength?: number;
+  maxLength?: number;
+  min?: number;
+  max?: number;
+}
+
+//validateの関数定義
+export function validate(validatableInput: Validatable) {
+  let isValid = true; //これの必要性は？
+  if (validatableInput.required) {
+    isValid = isValid && validatableInput.value.toString().trim().length !== 0;
+  }
+  if (
+    validatableInput.minLength != null &&
+    typeof validatableInput.value === 'string'
+  ) {
+    isValid =
+      isValid && validatableInput.value.length >= validatableInput.minLength;
+  }
+  if (
+    validatableInput.maxLength != null &&
+    typeof validatableInput.value === 'string'
+  ) {
+    isValid =
+      isValid && validatableInput.value.length <= validatableInput.maxLength;
+  }
+  if (
+    validatableInput.min != null &&
+    typeof validatableInput.value === 'number'
+  ) {
+    isValid = isValid && validatableInput.value >= validatableInput.min;
+  }
+  if (
+    validatableInput.max != null &&
+    typeof validatableInput.value === 'number'
+  ) {
+    isValid = isValid && validatableInput.value <= validatableInput.max;
+  }
+
+  return isValid;
+}
