+++ tsconfig.json
@@ -4,6 +4,7 @@
     // File Layout
     "rootDir": "./src",
     "outDir": "./dist",
+    // "outFile": "./dist/app.js",
 
     // Environment Settings
     // See also https://aka.ms/tsconfig/module
